refactor: move some functions into aws module
